# swagger_file.yaml
openapi: 3.0.0
info:
  title: Todo API
  description: API for managing todos
  version: 1.0.0
servers:
  - url: http://localhost:5000  # Update the port to match your Flask app
paths:
  /todos:  # Ensure the path matches your Flask endpoint
    get:
      summary: Get all todos
      responses:
        '200':
          description: A list of todo items
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    title:
                      type: string
                    completed:
                      type: boolean
    post:
      summary: Create a new todo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                completed:
                  type: boolean
      responses:
        '201':
          description: Todo created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  title:
                    type: string
                  completed:
                    type: boolean
  /todos/{id}:  # Ensure the path matches your Flask endpoint
    get:
      summary: Get a specific todo by ID
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the todo to fetch
          schema:
            type: integer
      responses:
        '200':
          description: Specific todo found
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  title:
                    type: string
                  completed:
                    type: boolean
        '404':
          description: Todo not found
    put:
      summary: Update an existing todo
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the todo to update
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                completed:
                  type: boolean
      responses:
        '200':
          description: Todo updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  title:
                    type: string
                  completed:
                    type: boolean
        '404':
          description: Todo not found
    delete:
      summary: Delete a todo
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the todo to delete
          schema:
            type: integer
      responses:
        '204':
          description: Todo deleted
        '404':
          description: Todo not found